+++ src/test/java/inshining/virtualstorage/metadata/service/FolderMetaDataServiceTest.java
@@ -241,4 +241,5 @@ public class FolderMetaDataServiceTest {
         //then
         assertFalse(result);
     }
+
 }
